restructure cashier auth components hierarchy
