+++ src/app/authentication/authentication-routing.module.ts
@@ -3,6 +3,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {AuthenticationComponent} from "./authentication.component";
 import {LoginPageComponent} from "./components/login-page/login-page.component";
 import {RegPageComponent} from "./components/reg-page/reg-page.component";
+import {CashierComponent} from "./components/cashier/cashier.component";
 
 const routes: Routes = [
   // {path: '', redirectTo: 'sign-in', pathMatch: 'full'},
@@ -11,6 +12,7 @@ const routes: Routes = [
     children: [
       {path: '', component: LoginPageComponent},
       {path: 'sign-up', component: RegPageComponent},
+      {path: 'cashier', component: CashierComponent},
     ],
   }];
