+++ src/app/authentication/components/cashier/cashier.component.html
