+++ src/app/cashier/cashier.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AuthComponent } from './components/auth/auth.component';
 import {CashierComponent} from "./cashier.component";
 import {CashierRoutingModule} from "./cashier-routing.module";
 import {ReactiveFormsModule} from "@angular/forms";
@@ -10,7 +9,6 @@ import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
 
 @NgModule({
   declarations: [
-    AuthComponent,
     CashierComponent
   ],
   imports: [
